fix(client): Bevy 0.19 runtime integration
- fix(client): Tonemapping::None to fix pink rendering (tonemapping_luts feature unavailable)
- feat(client): Console debug HUD with movement/input tracking via eprintln
- feat(client): WASD movement with smooth acceleration/deceleration (FreeCamera-style)
- feat(client): Proper terrain colors per type (Grass/Forest/Water/City/Desert/Polluted)
- feat(client): Multi-light lighting setup (GlobalAmbient + Directional + Point)
- refactor(client): Player mesh from box to capsule shape
- refactor(client): Player color from pink to teal
- fix(client): Invert W/S input for correct forward/backward movement
- fix(core): 27 compilation errors (imports, Copy derives, Option patterns, type casts, method syntax)
- fix(server): 16 compilation errors (scheduler module, serde imports, Clone on HexTileDbEntry, JSON serialization)
- fix(client): 42 compilation errors (lib.rs pub fn main, orphaned fn, Bevy 0.19 API mismatches)
- feat(server): Add scheduler/mod.rs with process_idle_gains
- feat(server): Add 7 missing world functions (handle_login, move_player, teleport_player, etc.)

Specs 001-004 compilation verified: 85 errors fixed across 3 crates.

+++ crates/idlecore-core/src/hex_tile.rs
@@ -45,7 +45,7 @@ impl HexTileData {
             center_x: x,
             center_y: y,
             terrain,
-            eco_rating: crate::terrain::eco_rating(&terrain),
+            eco_rating: crate::terrain::eco_rating(&terrain) as u32,
             owned_by: None,
             has_plant: false,
             has_pollution: false,
